fix: bug in tambah permintaan pengiriman

+++ silogistik2106639485/src/main/java/apap/ti/silogistik2106639485/repository/GudangDb.java
@@ -4,9 +4,7 @@ import apap.ti.silogistik2106639485.model.Gudang;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface GudangDb extends JpaRepository<Gudang, BigInteger> {
+public interface GudangDb extends JpaRepository<Gudang, Long> {
     
 }
